0032723: Draw Harness, ViewerTest - sloppy animation in WebGL
Fixed emscripten_async_call() calls to use requestAnimationFrame() instead of setTimeout().
ViewerTest_EventManager::handleViewRedraw() - fixed queuing new onWasmRedrawView()
before previous one has been processed (leading to multiplying of pending redraws executed at the same time).
RWGltf_TriangulationReader::readDracoBuffer - suppressed CLang compilation warnings.

+++ src/RWGltf/RWGltf_TriangulationReader.cxx
@@ -224,7 +224,7 @@ bool RWGltf_TriangulationReader::readDracoBuffer (const Handle(RWGltf_GltfLatePr
 
 #ifdef HAVE_DRACO
   std::vector<char> aReadData;
-  aReadData.resize (theGltfData.StreamLength);
+  aReadData.resize ((size_t )theGltfData.StreamLength);
   aSharedStream->read (aReadData.data(), (std::streamsize )theGltfData.StreamLength);
   if (!aSharedStream->good())
   {
